Changed finalStars
Unable to grab the finalStar count. Set variable outside of function to test and its grabbing that value, not the value of the finalStar. Without finalStar outside the function the modal does not come up at all. It only come up when correct data is inside it. So If I took out finalStar in the text in the modal the modal will come up with the final data for the moves and time(time only sec for now).